+++ source/rafcon/gui/helpers/state_machine.py
@@ -521,6 +521,8 @@ def refresh_selected_state_machine():
             logger.debug("Refresh of selected state machine canceled")
             return
 
+    library_manager.clean_loaded_libraries()
+    refresh_libraries()
     states_editor_ctrl.close_pages_for_specific_sm_id(selected_sm_id)
     state_machines_editor_ctrl.refresh_state_machine_by_id(selected_sm_id)
 
@@ -568,6 +570,7 @@ def refresh_all(force=False):
                 logger.debug("Refresh canceled")
                 return
 
+    library_manager.clean_loaded_libraries()
     refresh_libraries()
     states_editor_ctrl.close_all_pages()
     state_machines_editor_ctrl.refresh_all_state_machines()